requirements.txt
django==3.1.2
environs==8.0.0
python-telegram-bot==11.1.0
python-dotenv==0.14.0
phonenumbers==8.12.9
redis==3.5.3
